one more numpy op done

+++ src/test/cbackends/host/39.numpy/genNewOp.sh
@@ -7,26 +7,26 @@ set -o pipefail
 # Activate debug
 # set -x
 
-# if [ -f $1.cpp ]; then
-# 	echo "already generated, neglect"
-# 	sleep 1
-# else
+if [ -f $1.cpp ]; then
+	echo "already generated, neglect"
+	sleep 1
+else
 	sed 's/divide/'"$1"'/g' divide.cpp > $1.cpp
 	echo "change data please ..."
 	sleep 1
 	nvim "+normal 10G26|" $1.cpp
-# fi
+fi
 
-# if [ -f $1.py ]; then
-# 	echo "already generated, neglect"
-# 	sleep 1
-# else
+if [ -f $1.py ]; then
+	echo "already generated, neglect"
+	sleep 1
+else
 	sed 's/divide/'"$1"'/g' ./numpy/divide.py > ./numpy/$1.py
 	chmod +x ./numpy/$1.py
 	echo "change to the same data please ..."
 	sleep 1
 	nvim "+normal 6G14|" ./numpy/$1.py
-# fi
+fi
 
 echo "add compilation commands please ..."
 sleep 1
